+++ SDF/src/Structures/ShapeDiameterFunction/CSDF.h
@@ -9,20 +9,20 @@ namespace SDFStructures
 		CSDF();
 		~CSDF();
 
-		void ComputeValue(const std::vector<double> values, const std::vector<double> inverse_Yangles);
-		double CalcMedian(const std::vector<double> values, unsigned int size);
-		double CalcMean(const std::vector<double> values, unsigned int size);
-		double CalcStandardDeviation(const std::vector<double> values, unsigned int size);
-		void Smooth(const std::vector<double> values, const std::vector<double> weights);
+		void ComputeValue(const std::vector<float> values, const std::vector<float> inverse_Yangles);
+		float CalcMedian(const std::vector<float> values, unsigned int size);
+		float CalcMean(const std::vector<float> values, unsigned int size);
+		float CalcStandardDeviation(const std::vector<float> values, unsigned int size);
+		void Smooth(const std::vector<float> values, const std::vector<float> weights);
 
 
-		void Normalize1(double min, double max, double alfa);
-		void Normalize2(double min, double max, double alfa);
+		void Normalize1(float min, float max, float alfa);
+		void Normalize2(float min, float max, float alfa);
 
 
-		double value;
-		double smoothed;
-		double normalized1;
-		double normalized2;
+		float value;
+		float smoothed;
+		float normalized1;
+		float normalized2;
 	};
 }
